+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.17.1-alpha.1
+- [#651] Change the rewind behavior. Now if the rewind is used while the sync process is in progress then an exception is thrown.
 - [#616] Download Stream generates too many updates on the main thread
 
 > **WARNING**: Notifications from SDK are no longer delivered on main thread.